requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Pillow>=9.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
PyPDF2==3.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
cryptography==41.0.4
schedule==1.2.0
APScheduler==3.10.4
openpyxl==3.1.2
reportlab==4.0.4
qrcode==7.4.2
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
psutil==5.9.5
pytest==7.4.0
pytest-flask==1.2.0 